multi: move GenChallengeNUMS
To avoid a circular package dependency between the address and proof
package, we move the GenChallengeNUMS function into the assets package.
That function is the only reference the proof package has to the address
package, so the move removes that dependency.
